+++ src/compiler/compile/nodes/Binding.ts
@@ -3,7 +3,7 @@ import get_object from '../utils/get_object';
 import Expression from './shared/Expression';
 import Component from '../Component';
 import TemplateScope from './shared/TemplateScope';
-import {dimensions} from '../../utils/patterns';
+import { regex_dimensions } from '../../utils/patterns';
 import { Node as ESTreeNode } from 'estree';
 import { TemplateNode } from '../../interfaces';
 import Element from './Element';
@@ -11,6 +11,7 @@ import InlineComponent from './InlineComponent';
 import Window from './Window';
 import { clone } from '../../utils/clone';
 import compiler_errors from '../compiler_errors';
+import compiler_warnings from '../compiler_warnings';
 
 // TODO this should live in a specific binding
 const read_only_media_attributes = new Set([
@@ -47,6 +48,7 @@ export default class Binding extends Node {
 		const { name } = get_object(this.expression.node);
 
 		this.is_contextual = Array.from(this.expression.references).some(name => scope.names.has(name));
+		if (this.is_contextual) this.validate_binding_rest_properties(scope);
 
 		// make sure we track this as a mutable ref
 		if (scope.is_let(name)) {
@@ -86,7 +88,7 @@ export default class Binding extends Node {
 		const type = parent.get_static_attribute_value('type');
 
 		this.is_readonly =
-			dimensions.test(this.name) ||
+			regex_dimensions.test(this.name) ||
 			(isElement(parent) &&
 				((parent.is_media_node() && read_only_media_attributes.has(this.name)) ||
 					(parent.name === 'input' && type === 'file')) /* TODO others? */);
@@ -94,6 +96,18 @@ export default class Binding extends Node {
 
 	is_readonly_media_attribute() {
 		return read_only_media_attributes.has(this.name);
+	}
+
+	validate_binding_rest_properties(scope: TemplateScope) {
+		this.expression.references.forEach(name => {
+			const each_block = scope.get_owner(name);
+			if (each_block && each_block.type === 'EachBlock') {
+				const rest_node = each_block.context_rest_properties.get(name);
+				if (rest_node) {
+					this.component.warn(rest_node as any, compiler_warnings.invalid_rest_eachblock_binding(name));
+				}
+			}
+		});
 	}
 }