+++ src/compiler/compile/nodes/shared/Expression.ts
@@ -21,6 +21,8 @@ import compiler_errors from '../../compiler_errors';
 
 type Owner = INode;
 
+const regex_contains_term_function_expression = /FunctionExpression/;
+
 export default class Expression {
 	type: 'Expression' = 'Expression';
 	component: Component;
@@ -72,7 +74,7 @@ export default class Expression {
 					scope = map.get(node);
 				}
 
-				if (!function_expression && /FunctionExpression/.test(node.type)) {
+				if (!function_expression && regex_contains_term_function_expression.test(node.type)) {
 					function_expression = node;
 				}
 
@@ -126,6 +128,7 @@ export default class Expression {
 						deep = node.left.type === 'MemberExpression';
 						names = extract_names(deep ? get_object(node.left) : node.left);
 					} else if (node.type === 'UpdateExpression') {
+                        deep = node.argument.type === 'MemberExpression';
 						names = extract_names(get_object(node.argument));
 					}
 				}
@@ -147,7 +150,26 @@ export default class Expression {
 							component.add_reference(node, name);
 
 							const variable = component.var_lookup.get(name);
-							if (variable) variable[deep ? 'mutated' : 'reassigned'] = true;
+
+							if (variable) {
+								variable[deep ? 'mutated' : 'reassigned'] = true;
+							}
+
+							const declaration: any = scope.find_owner(name)?.declarations.get(name);
+
+							if (declaration) {
+								if (declaration.kind === 'const' && !deep) {
+									component.error(node, {
+										code: 'assignment-to-const',
+										message: 'You are assigning to a const'
+									});
+								}
+							} else if (variable && variable.writable === false && !deep) {
+								component.error(node, {
+									code: 'assignment-to-const',
+									message: 'You are assigning to a const'
+								});
+							}
 						}
 					});
 				}
@@ -252,21 +274,55 @@ export default class Expression {
 					);
 
 					const declaration = b`const ${id} = ${node}`;
+					const extract_functions = () => {
+						const deps = Array.from(contextual_dependencies);
+						const function_expression = node as FunctionExpression;
+
+						const has_args = function_expression.params.length > 0;
+						function_expression.params = [
+							...deps.map(name => ({ type: 'Identifier', name } as Identifier)),
+							...function_expression.params
+						];
+
+						const context_args = deps.map(name => block.renderer.reference(name, ctx));
+
+						component.partly_hoisted.push(declaration);
+
+						block.renderer.add_to_context(id.name);
+						const callee = block.renderer.reference(id);
+
+						this.replace(id as any);
+
+						const func_declaration = has_args
+							? b`function ${id}(...args) {
+								return ${callee}(${context_args}, ...args);
+							}`
+							: b`function ${id}() {
+								return ${callee}(${context_args});
+							}`;
+						return { deps, func_declaration };
+					};
 
 					if (owner.type === 'ConstTag') {
-						let child_scope = scope;
-						walk(node, {
-							enter(node: Node, parent: any) {
-								if (map.has(node)) child_scope = map.get(node);
-								if (node.type === 'Identifier' && is_reference(node, parent)) {
-									if (child_scope.has(node.name)) return;
-									this.replace(block.renderer.reference(node, ctx));
+						// we need a combo block/init recipe
+						if (contextual_dependencies.size === 0) {
+							let child_scope = scope;
+							walk(node, {
+								enter(node: Node, parent: any) {
+									if (map.has(node)) child_scope = map.get(node);
+									if (node.type === 'Identifier' && is_reference(node, parent)) {
+										if (child_scope.has(node.name)) return;
+										this.replace(block.renderer.reference(node, ctx));
+									}
+								},
+								leave(node: Node) {
+									if (map.has(node)) child_scope = child_scope.parent;
 								}
-							},
-							leave(node: Node) {
-								if (map.has(node)) child_scope = child_scope.parent;
-							}
-						});
+							});
+						} else {
+							const { func_declaration } = extract_functions();
+							this.replace(func_declaration[0]);
+						}
 					} else if (dependencies.size === 0 && contextual_dependencies.size === 0) {
 						// we can hoist this out of the component completely
 						component.fully_hoisted.push(declaration);
@@ -287,31 +343,7 @@ export default class Expression {
 						this.replace(block.renderer.reference(id));
 					} else {
 						// we need a combo block/init recipe
-						const deps = Array.from(contextual_dependencies);
-						const function_expression = node as FunctionExpression;
-
-						const has_args = function_expression.params.length > 0;
-						function_expression.params = [
-							...deps.map(name => ({ type: 'Identifier', name } as Identifier)),
-							...function_expression.params
-						];
-
-						const context_args = deps.map(name => block.renderer.reference(name));
-
-						component.partly_hoisted.push(declaration);
-
-						block.renderer.add_to_context(id.name);
-						const callee = block.renderer.reference(id);
-
-						this.replace(id as any);
-
-						const func_declaration = has_args
-							? b`function ${id}(...args) {
-								return ${callee}(${context_args}, ...args);
-							}`
-							: b`function ${id}() {
-								return ${callee}(${context_args});
-							}`;
+						const { deps, func_declaration } = extract_functions();
 
 						if (owner.type === 'Attribute' && owner.parent.name === 'slot') {
 							const dep_scopes = new Set<INode>(deps.map(name => template_scope.get_owner(name)));
@@ -323,7 +355,7 @@ export default class Expression {
 
 							const func_expression = func_declaration[0];
 
-							if (node.type === 'InlineComponent') {
+							if (node.type === 'InlineComponent' || node.type === 'SlotTemplate') {
 								// <Comp let:data />
 								this.replace(func_expression);
 							} else {