+++ src/compiler/compile/utils/hash.ts
@@ -1,6 +1,9 @@
 // https://github.com/darkskyapp/string-hash/blob/master/index.js
+
+const regex_return_characters = /\r/g;
+
 export default function hash(str: string): string {
-	str = str.replace(/\r/g, '');
+	str = str.replace(regex_return_characters, '');
 	let hash = 5381;
 	let i = str.length;