+++ src/runtime/index.ts
@@ -13,5 +13,5 @@ export {
 	createEventDispatcher,
 	SvelteComponentDev as SvelteComponent,
 	SvelteComponentTyped
-	// additional exports added through post-typegen.js
+	// additional exports added through generate-type-definitions.js
 } from 'svelte/internal';