+++ src/runtime/internal/await_block.ts
@@ -2,11 +2,36 @@ import { is_promise } from './utils';
 import { check_outros, group_outros, transition_in, transition_out } from './transitions';
 import { flush } from './scheduler';
 import { get_current_component, set_current_component } from './lifecycle';
+import { Fragment, FragmentFactory } from './types';
 
-export function handle_promise(promise, info) {
+interface PromiseInfo<T> {
+	ctx: null | any;
+	// unique object instance as a key to compare different promises
+	token: {},
+	hasCatch: boolean,
+	pending: FragmentFactory,
+	then: FragmentFactory,
+	catch: FragmentFactory,
+	// ctx index for resolved value and rejected error
+	value: number,
+	error: number,
+	// resolved value or rejected error
+	resolved?: T,
+	// the current factory function for creating the fragment
+	current: FragmentFactory | null,
+	// the current fragment
+	block: Fragment | null,
+	// tuple of the pending, then, catch fragment
+	blocks: [null | Fragment, null | Fragment, null | Fragment];
+	// DOM elements to mount and anchor on for the {#await} block
+	mount: () => HTMLElement;
+	anchor: HTMLElement;
+}
+
+export function handle_promise<T>(promise: Promise<T>, info: PromiseInfo<T>) {
 	const token = info.token = {};
 
-	function update(type, index, key?, value?) {
+	function update(type: FragmentFactory, index: 0 | 1 | 2, key?: number, value?) {
 		if (info.token !== token) return;
 
 		info.resolved = value;