+++ test/validator/samples/a11y-aria-proptypes-integer/input.svelte
@@ -1,0 +1,7 @@
+<div aria-level="yes" />
+<div aria-level="no" />
+<div aria-level={`abc`} />
+<div aria-level={true} />
+<div aria-level />
+<div aria-level={"false"} />
+<div aria-level={!"false"} />