+++ test/validator/samples/css-invalid-combinator-selector-3/input.svelte
@@ -1,0 +1,9 @@
+<p><span /></p>
+
+<style>
+	@media screen {
+		> span {
+			color: red;
+		}
+	}
+</style>